+++ R/gsDesignTune-package.R
@@ -6,5 +6,6 @@
 #' @importFrom progressr progressor with_progress
 #' @importFrom R6 R6Class
 #' @importFrom rlang .data
+#' @importFrom tinytable format_tt style_tt theme_html tt
 ## usethis namespace: end
 NULL